+++ breeze-core/src/main/kotlin/com/windea/breezeframework/core/extensions/CharExtensions.kt
@@ -3,7 +3,7 @@
 
 package com.windea.breezeframework.core.extensions
 
-//region operator overrides
+//region operator override extensions
 /**@see com.windea.breezeframework.core.extensions.repeat */
 operator fun Char.times(n: Int): String = this.repeat(n)
 //endregion