+++ breeze-time/src/main/kotlin/com/windea/breezeframework/time/extensions/TemporalAmountExtensions.kt
@@ -6,7 +6,7 @@ package com.windea.breezeframework.time.extensions
 import java.time.*
 import java.time.temporal.*
 
-//region operator overrides
+//region operator override extensions
 /**@see java.time.Duration.negated*/
 operator fun Duration.unaryMinus(): Duration = this.negated()