+++ CHANGELOG.md
@@ -257,10 +257,8 @@
 * [X] breeze-core 优化字符串的特定类型操作方法以及集合的深操作方法。
 * [X] breeze-core 移除一些认为是冗余的扩展。
 * [X] breeze-core 添加`Array.swap`、`MutableList.swap`扩展。
+* [X] breeze-core&breeze-reflect 添加`Any.isInstanceOf`扩展。
 * [ ] breeze-core 让集合的`deepQuery`方法当特定的占位符不匹配类型时，回调为普通字符串。
-
-# 待办
-
 * [ ] breeze-dsl 实现`YamlDsl`。（不推荐使用，但是作为规范）
 * [ ] breeze-dsl 优化：对于表示转化的元素，可以通过`"a"(...) fromTo "b"`的语法构建。
 * [ ] breeze-serialization 提供yaml的序列化实现。